Update emphasis patch to use RBJ bell.
It's alive! This looks much more promising on measurements. May want
to tweak gain and resonance range after a bit of hands-on testing.

There does seem to be a bit of high-frequency roll-off, even at flat
settings. It's not really audible, but it also shouldn't be necessary
especially running at 96k. Odd.

See if it's also there when I implement shelves?

I'm not noticing any zipper noise, which is great as there is
currently no parameter smoothing. Might be more obvious on purer input
sources, or with heavy modulation?

Implementation could be slightly more efficient as we are using two
instances of Biquad coefficient calculator, but the values are nearly
the same. At this point, maybe don't worry about it. But the option is
there to make an "emphasis biquad" that is designed for such things.